// src/types/ShippingManagement.ts
export type Shipping = {
  id: string;
  name: string;
  father: string;
  mothersFather: string;
  mothersGrandfather: string;
  grandmothersGrandfather: string;
  matingDate: string;
  expectedBirthDate: string;
  birthDate: string;
  auctionDate: string;
  weight: number;
  daysOld: number;
  sex: string;
  price: number;
  buyer: string;
  memo: string;
};